# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.1.1-oas3",
    "title": "Teste nodeJS locadora",
    "description": "Faz o cadastro de filmes e pessoas, lista filmes, aluga filmes e faz o calculo do valor da devolução destes mesmos filmes.",
    "termsOfService": "http://localhost:3333/",
    "contact": {
      "name": "Github",
      "url": "https://github.com/GuilhermeSSilva/teste-nodejs",
      "email": "guilherme_rgcosta@hotmail.com"
    }
  },
  "tags": [
    {
      "name": "Movies"
    },
    {
      "name": "Users"
    },
    {
      "name": "Rent"
    }
  ],
  "paths": {
    "/movies": {
      "post": {
        "tags": [
          "Movies"
        ],
        "summary": "Cadastra files",
        "description": "Cadastra um novo filme",
        "parameters": [
          {
            "name": "movie",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Star Wars"
                },
                "director": {
                  "type": "string",
                  "example": "George Lucas"
                },
                "rentValue": {
                  "type": "number",
                  "example": 10
                },
                "numberCopies": {
                  "type": "number",
                  "example": 5
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Filme registrado com sucesso."
          },
          "400": {
            "description": "Por favor preencha todos os campos"
          },
          "500": {
            "description": "Não foi possível cadastrar o filme no momento, por favor tente novamente mais tarde."
          }
        }
      },
      "get": {
        "tags": [
          "Movies"
        ],
        "summary": "Lista todos os filmes.",
        "description": "Lista todos os filmes.",
        "responses": {
          "200": {
            "description": "Ok",
            "schema": {
              "$ref": "#/definitions/movies"
            }
          },
          "500": {
            "description": "Não conseguimos listar os filmes no momento, por favor tente novamente mais tarde"
          }
        }
      }
    },
    "/rent": {
      "post": {
        "tags": [
          "Rent"
        ],
        "summary": "Aluga filmes.",
        "description": "Cria um registro de alugado na tabela rent.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nameUser": {
                  "type": "string",
                  "example": "Guilherme"
                },
                "cpfUser": {
                  "type": "string",
                  "example": "12345678901"
                },
                "idMovies": {
                  "type": "number",
                  "example": "[1,2]"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/rent"
            }
          },
          "404": {
            "description": "Desculpe mas estes filmes já foram alugados"
          },
          "500": {
            "description": "Desculpa mas não conseguimos alugar filmes no momento"
          }
        }
      }
    },
    "/user": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Cadastra usuário",
        "description": "Cadastra um usuário",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Guilherme"
                },
                "cpf": {
                  "type": "string",
                  "example": 12345678901
                },
                "telephone": {
                  "type": "string",
                  "example": 53981072737
                },
                "address": {
                  "type": "string",
                  "example": "Rua Leopoldo Hartel"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Usuário adicionado com sucesso."
          },
          "400": {
            "description": "Por favor preencha todos os campos"
          },
          "500": {
            "description": "Não foi possível adicionar o usuário no momento, por favor tente mais tarde."
          }
        }
      }
    },
    "/returnMovie": {
      "post": {
        "tags": [
          "Rent"
        ],
        "summary": "Devolve os filmes alugados e mostra o valor a ser pago.",
        "description": "Devolve os filmes alugados e mostra o valor a ser pago.",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nameUser": {
                  "type": "string",
                  "example": "Guilherme Silva"
                },
                "cpfUser": {
                  "type": "string",
                  "example": 12345678901
                },
                "idMovies": {
                  "type": "string",
                  "example": "[1,2]"
                }
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/returnMovie"
            }
          },
          "400": {
            "description": "Não há filmes a serem devolvidos"
          },
          "500": {
            "description": "Houve um problema no nosso servidor, por favor tente novamente mais tarde"
          }
        }
      }
    }
  },
  "definitions": {
    "returnMovie": {
      "type": "object",
      "properties": {
        "value": {
          "type": "number",
          "example": 20
        }
      }
    },
    "movies": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "title": {
          "type": "string",
          "example": "Star Wars"
        },
        "director": {
          "type": "string",
          "example": "George Lucas"
        },
        "rentValue": {
          "type": "number",
          "example": 5
        },
        "numberCopies": {
          "type": "number",
          "example": 10
        }
      }
    },
    "rent": {
      "type": "object",
      "properties": {
        "namesMovies": {
          "type": "string",
          "example": "['Duro de matar', 'Deus não está morto']"
        },
        "value": {
          "type": "number",
          "example": 15
        },
        "returnDate": {
          "type": "string",
          "example": "2021-09-06T00:00:00.000Z"
        },
        "returnTime": {
          "type": "string",
          "example": "03:13:44.114"
        }
      }
    }
  }
}